nutrirx/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
google-genai>=1.0.0
google-cloud-vision>=3.7.0
httpx==0.27.0
pydantic==2.7.1
aiosqlite==0.20.0
pillow==10.3.0
sse-starlette==2.1.0
